+++ packages/@expo/server/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🎉 New features
 
+- Add new Runtime Server API with `environment()`, `origin()`, `runTask()`, `deferTask()`, and `StatusError` ([#39925](https://github.com/expo/expo/pull/39925) by [@kitten](https://github.com/kitten))
+
 ### 🐛 Bug fixes
 
 - Use Headers to copy response headers (prevent `set-cookie` lost) ([#39846](https://github.com/expo/expo/pull/39846) by [@LevanKvirkvelia](https://github.com/LevanKvirkvelia))